+++ src/main/java/net/jeqo/bloons/Bloons.java
@@ -10,10 +10,10 @@ import net.jeqo.bloons.listeners.*;
 import net.jeqo.bloons.listeners.multipart.MultipartBalloonPlayerJoinListener;
 import net.jeqo.bloons.listeners.multipart.MultipartBalloonPlayerLeaveListener;
 import net.jeqo.bloons.listeners.single.SingleBalloonPlayerListener;
-import net.jeqo.bloons.utils.Languages;
-import net.jeqo.bloons.utils.UpdateChecker;
+import net.jeqo.bloons.message.Languages;
+import net.jeqo.bloons.health.UpdateChecker;
 import net.jeqo.bloons.logger.Logger;
-import net.jeqo.bloons.utils.Metrics;
+import net.jeqo.bloons.health.Metrics;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.util.HashMap;
