+++ src/main/java/net/jeqo/bloons/balloon/single/SingleBalloon.java
@@ -7,9 +7,9 @@ import net.jeqo.bloons.configuration.BalloonConfiguration;
 import net.jeqo.bloons.events.balloon.single.SingleBalloonEquipEvent;
 import net.jeqo.bloons.events.balloon.single.SingleBalloonForceUnequipEvent;
 import net.jeqo.bloons.logger.Logger;
-import net.jeqo.bloons.utils.Languages;
-import net.jeqo.bloons.utils.management.SingleBalloonManagement;
-import net.jeqo.bloons.utils.Color;
+import net.jeqo.bloons.message.Languages;
+import net.jeqo.bloons.management.SingleBalloonManagement;
+import net.jeqo.bloons.colors.Color;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Location;
 import org.bukkit.Material;
