+++ src/main/java/net/jeqo/bloons/colors/ColorCodeConverter.java
@@ -1,4 +1,4 @@
-package net.jeqo.bloons.utils;
+package net.jeqo.bloons.colors;
 
 import java.util.HashMap;
 import java.util.Map;
