+++ src/main/java/net/jeqo/bloons/commands/CommandForceUnequip.java
@@ -7,10 +7,10 @@ import net.jeqo.bloons.commands.manager.Command;
 import net.jeqo.bloons.commands.manager.types.CommandPermission;
 import net.jeqo.bloons.events.balloon.multipart.MultipartBalloonUnequipEvent;
 import net.jeqo.bloons.events.balloon.single.SingleBalloonUnequipEvent;
-import net.jeqo.bloons.utils.Languages;
-import net.jeqo.bloons.utils.management.SingleBalloonManagement;
-import net.jeqo.bloons.utils.MessageTranslations;
-import net.jeqo.bloons.utils.management.MultipartBalloonManagement;
+import net.jeqo.bloons.message.Languages;
+import net.jeqo.bloons.management.SingleBalloonManagement;
+import net.jeqo.bloons.message.MessageTranslations;
+import net.jeqo.bloons.management.MultipartBalloonManagement;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
