+++ src/main/java/net/jeqo/bloons/commands/CommandReload.java
@@ -4,8 +4,8 @@ import net.jeqo.bloons.Bloons;
 import net.jeqo.bloons.commands.manager.Command;
 import net.jeqo.bloons.commands.manager.types.CommandPermission;
 import net.jeqo.bloons.events.general.BloonsConfigReloadEvent;
-import net.jeqo.bloons.utils.Languages;
-import net.jeqo.bloons.utils.MessageTranslations;
+import net.jeqo.bloons.message.Languages;
+import net.jeqo.bloons.message.MessageTranslations;
 import net.kyori.adventure.text.Component;
 import org.bukkit.command.CommandSender;
 import org.bukkit.plugin.java.JavaPlugin;
