+++ src/main/java/net/jeqo/bloons/commands/utils/ErrorHandling.java
@@ -2,7 +2,7 @@ package net.jeqo.bloons.commands.utils;
 
 import net.jeqo.bloons.Bloons;
 import net.jeqo.bloons.configuration.PluginConfiguration;
-import net.jeqo.bloons.utils.MessageTranslations;
+import net.jeqo.bloons.message.MessageTranslations;
 import net.kyori.adventure.text.Component;
 import org.bukkit.command.CommandSender;
