+++ src/main/java/net/jeqo/bloons/gui/GUIClickableItem.java
@@ -1,6 +1,6 @@
 package net.jeqo.bloons.gui;
 
-import net.jeqo.bloons.utils.NBTItem;
+import net.jeqo.bloons.item.NBTItem;
 import org.bukkit.event.inventory.InventoryClickEvent;
 
 import java.util.HashMap;
