+++ src/main/java/net/jeqo/bloons/gui/menus/BalloonMenu.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 import lombok.Setter;
 import net.jeqo.bloons.Bloons;
 import net.jeqo.bloons.logger.Logger;
-import net.jeqo.bloons.utils.Color;
-import net.jeqo.bloons.utils.Languages;
-import net.jeqo.bloons.utils.MessageTranslations;
+import net.jeqo.bloons.colors.Color;
+import net.jeqo.bloons.message.Languages;
+import net.jeqo.bloons.message.MessageTranslations;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
