+++ src/main/java/net/jeqo/bloons/listeners/multipart/MultipartBalloonPlayerListener.java
@@ -6,8 +6,8 @@ import net.jeqo.bloons.balloon.multipart.balloon.MultipartBalloon;
 import net.jeqo.bloons.balloon.multipart.balloon.MultipartBalloonBuilder;
 import net.jeqo.bloons.events.balloon.multipart.*;
 import net.jeqo.bloons.events.balloon.single.SingleBalloonForceUnequipEvent;
-import net.jeqo.bloons.utils.management.MultipartBalloonManagement;
-import net.jeqo.bloons.utils.management.SingleBalloonManagement;
+import net.jeqo.bloons.management.MultipartBalloonManagement;
+import net.jeqo.bloons.management.SingleBalloonManagement;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.PlayerDeathEvent;
