+++ src/main/java/net/jeqo/bloons/listeners/single/SingleBalloonPlayerListener.java
@@ -4,9 +4,9 @@ import net.jeqo.bloons.Bloons;
 import net.jeqo.bloons.balloon.single.SingleBalloon;
 import net.jeqo.bloons.events.balloon.single.SingleBalloonForceUnequipEvent;
 import net.jeqo.bloons.events.balloon.single.SingleBalloonStoreEvent;
-import net.jeqo.bloons.utils.UpdateChecker;
+import net.jeqo.bloons.health.UpdateChecker;
 import net.jeqo.bloons.logger.Logger;
-import net.jeqo.bloons.utils.management.SingleBalloonManagement;
+import net.jeqo.bloons.management.SingleBalloonManagement;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
