+++ src/main/java/net/jeqo/bloons/logger/Logger.java
@@ -3,8 +3,8 @@ package net.jeqo.bloons.logger;
 import net.jeqo.bloons.Bloons;
 import net.jeqo.bloons.configuration.ConfigConfiguration;
 import net.jeqo.bloons.configuration.PluginConfiguration;
-import net.jeqo.bloons.utils.Languages;
-import net.jeqo.bloons.utils.MessageTranslations;
+import net.jeqo.bloons.message.Languages;
+import net.jeqo.bloons.message.MessageTranslations;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
